added support for dns resolver in mqtt module and config Ubidots broker

+++ src/network/mqtt/MQTT_utils.hpp
@@ -2,5 +2,5 @@
 
 struct level_sensor{
     uint8_t unit;
-    uint8_t value;
+    uint32_t value;
 };
